separate merging stuff from classification, filter
