+++ q2_annotate/kraken2/filter.py
@@ -562,149 +562,3 @@ def _align_single_output_with_report(
 
     output_fp = os.path.join(output_dir_fmt.path, f"{sample_id}.output.txt")
     output_df.to_csv(output_fp, sep='\t', header=None, index=None)
-
-
-def _merge_trees(
-    first: tuple[TreeNode | None, TreeNode | None],
-    second: tuple[TreeNode | None, TreeNode | None]
-) -> tuple[TreeNode | None, TreeNode | None]:
-    '''
-    Merges two trees each representing a kraken2 report into a single tree.
-    The number of reads assigned to each node are summed where nodes overlap,
-    and new nodes are inserted into the tree where they don't. The proportions
-    of assigned and covered reads are then updated in a final passover.
-
-    Parameters
-    ----------
-    first : tuple[TreeNode | None, TreeNode | None]
-        The first report tree, where the first node in the tuple represents the
-        tree and the second node represents an optional unclassified node.
-    second : tuple[TreeNode | None, TreeNode | None]
-        The second report tree, where the first node in the tuple represents the
-        tree and the second node represents an optional unclassified node.
-
-    Returns
-    -------
-    tuple[TreeNode | None, TreeNode | None]
-        The merged tree.
-    '''
-    first_tree, first_unclassified_node = first
-    second_tree, second_unclassified_node = second
-
-    # merge trees (with respect to `n_frags_assigned`, `n_frags_covered`)
-    if first_tree is None and second_tree is None:
-        merged_tree = None
-    elif first_tree is None:
-        merged_tree = second_tree
-    elif second_tree is None:
-        merged_tree = first_tree
-    else:
-        for node in first_tree.levelorder():
-            match = _find_node(second_tree, node)
-            if match is not None:
-                match._kraken_data['n_frags_assigned'] += \
-                    node._kraken_data['n_frags_assigned']
-                match._kraken_data['n_frags_covered'] += \
-                    node._kraken_data['n_frags_assigned']
-            else:
-                parent = _find_node(second_tree, node.parent)
-                new_node = TreeNode()
-                new_node._kraken_data = node._kraken_data
-                new_node._kraken_data['n_frags_covered'] = \
-                    new_node._kraken_data['n_frags_assigned']
-                parent.append(new_node)
-                match = new_node
-
-            # keep `n_frags_covered` up to date
-            for ancestor in match.ancestors():
-                ancestor._kraken_data['n_frags_covered'] += \
-                    node._kraken_data['n_frags_assigned']
-
-        merged_tree = second_tree
-
-    unclassified_node = _merge_unclassified_nodes(
-        first_unclassified_node, second_unclassified_node
-    )
-
-    # final passover to update `perc_frags_covered`
-    if merged_tree is not None:
-        total_reads = merged_tree._kraken_data['n_frags_covered']
-        if unclassified_node is not None:
-            total_reads += unclassified_node._kraken_data['n_frags_covered']
-
-        for node in merged_tree.traverse():
-            node._kraken_data['perc_frags_covered'] = round(
-                (node._kraken_data['n_frags_covered'] / total_reads) * 100, 2
-            )
-
-    return merged_tree, unclassified_node
-
-
-def _find_node(tree: TreeNode, node: TreeNode) -> TreeNode | None:
-    '''
-    Searches for `node` in `tree`, returns a match if there is one, otherwise
-    None. Matches are determined based on the kraken2-assigned taxon id.
-
-    Parameters
-    ----------
-    tree : skbio.TreeNode
-        The tree in which to search.
-    node : skbio.TreeNode
-        The node for which to search.
-
-    Returns
-    -------
-    skbio.TreeNode | None
-        The matching node if one was found, otherwise None.
-
-    Raises
-    ------
-    ValueError
-        If more than one matching node is found. Taxon ids should be unique
-        within a tree.
-    '''
-    def _match_by_taxon_id(n):
-        return n._kraken_data['taxon_id'] == node._kraken_data['taxon_id']
-
-    matches = list(tree.find_by_func(_match_by_taxon_id))
-
-    if len(matches) > 1:
-        raise ValueError('Did not expect more than one taxon id match.')
-    elif len(matches) == 1:
-        return matches[0]
-    else:
-        return None
-
-
-def _merge_unclassified_nodes(
-    first: TreeNode | None, second: TreeNode | None
-) -> TreeNode | None:
-    '''
-    Merges two TreeNodes representing unclassified nodes from kraken2 reports.
-
-    Parameters
-    ----------
-    first : TreeNode | None
-        The first unclassified node, or None if no such node exists.
-    second : TreeNode | None
-        The second unclassified node, or None if no such node exists.
-
-    Returns
-    -------
-    TreeNode | None
-        The merged unclassified node, or None if both inputs are None.
-    '''
-    if first is None and second is None:
-        unclassified_node = None
-    elif first is None:
-        unclassified_node = second
-    elif second is None:
-        unclassified_node = first
-    else:
-        unclassified_node = second
-        unclassified_node._kraken_data['n_frags_assigned'] += \
-            first._kraken_data['n_frags_assigned']
-        unclassified_node._kraken_data['n_frags_covered'] += \
-            first._kraken_data['n_frags_covered']
-
-    return unclassified_node
